Move model from db package to models package as it's also serialised to the client
