+++ common/src/main/scala/models/PlatformCount.scala
@@ -1,4 +1,4 @@
-package db
+package models
 
 import play.api.libs.json.{Format, Json}
