+++ common/src/test/scala/db/RegistrationServiceTest.scala
@@ -6,7 +6,7 @@ import doobie.implicits._
 import cats.effect.IO
 import cats.implicits._
 import doobie.util.transactor.Transactor
-import models.{Android, iOS}
+import models.{Android, PlatformCount, iOS}
 import org.specs2.specification.BeforeAll
 import fs2.Stream
 import org.specs2.concurrent.ExecutionEnv
